pyproject.toml
[project]
name = "mcp-bear"
version = "1.0.0"
description = "MCP server for accessing Bear Notes"
authors = [{ name = "Bora Gonul", email = "me@boragonul.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.1.2",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-bear = "mcp_bear.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_bear"]
